HSEARCH-3490 Move all Elasticsearch dialect implementations to a single package
We do not need each of them to have its own package, since the related
classes (work factory, ...) are implemented in completely different
packages, and thus we'll never have a related class in the
org.hibernate.search.backend.elasticsearch.dialect.impl.esXX package.
